requirements.txt
pymysql
python-dotenv
cryptography
streamlit
pandas